sort handlers into PartialHandlers and Middleware
