+++ src/SignIn/Program.cs
@@ -8,8 +8,10 @@ namespace SignIn
         {
             AuthorizationHelper.SetupPermissions();
 
+            new Middleware().Register();
             new CommitHooks().Register();
             new MainHandlers().Register();
+            new PartialHandlers().Register();
             new BlenderMapping().Register();
         }
     }
