+++ src/SignIn/SignIn.csproj
@@ -69,6 +69,8 @@
     <Compile Include="Api\CommitHooks.cs" />
     <Compile Include="Api\MainHandlers.cs" />
     <Compile Include="Api\BlenderMapping.cs" />
+    <Compile Include="Api\Middleware.cs" />
+    <Compile Include="Api\PartialHandlers.cs" />
     <Compile Include="Helpers\AuthorizationHelper.cs" />
     <Compile Include="Helpers\MailSettingsHelper.cs" />
     <Compile Include="Helpers\SignInJwt.cs" />
